+++ src/renderer/components/Tables/Cells.tsx
@@ -22,10 +22,15 @@ import { Language, Phrase } from 'localisation/types';
 import { Button } from '../Button/Button';
 import { Tooltip } from '../Tooltip/Tooltip';
 import { getLocalePhrase } from 'localisation/translations';
-import { LockKeyhole, LockOpen } from 'lucide-react';
-import StateManager from 'renderer/StateManager';
-import { MutableRefObject } from 'react';
+import {
+  LockKeyhole,
+  LockOpen,
+  MessageSquare,
+  MessageSquareMore,
+} from 'lucide-react';
+import { Dispatch, SetStateAction } from 'react';
 import { dungeonAffixesById } from 'main/constants';
+import TagDialog from 'renderer/TagDialog';
 
 export const populateResultCell = (
   info: CellContext<RendererVideo, unknown>,
@@ -66,7 +71,7 @@ export const populateMapCell = (info: CellContext<RendererVideo, unknown>) => {
 
 export const populateDateCell = (info: CellContext<RendererVideo, unknown>) => {
   const date = info.getValue() as Date;
-  return dateToHumanReadable(date);
+  return <div className="truncate">{dateToHumanReadable(date)}</div>;
 };
 
 export const populateActivityCell = (
@@ -90,11 +95,11 @@ export const populateActivityCell = (
 export const populateDetailsCell = (
   ctx: CellContext<RendererVideo, unknown>,
   appState: AppState,
-  stateManager: MutableRefObject<StateManager>,
+  setVideoState: Dispatch<SetStateAction<RendererVideo[]>>,
 ) => {
   const video = ctx.getValue() as RendererVideo;
   const { language, cloudStatus } = appState;
-  const { del } = cloudStatus;
+  const { write, del } = cloudStatus;
 
   const renderProtectedIcon = () => {
     // If any videos in our selection are not protected, then the button's
@@ -121,13 +126,29 @@ export const populateDetailsCell = (
 
     const toggleProtected = (e: React.MouseEvent<HTMLButtonElement>) => {
       stopPropagation(e);
-      stateManager.current.setProtected(lock, toProtect);
 
       window.electron.ipcRenderer.sendMessage('videoButton', [
         'protect',
         lock,
         toProtect,
       ]);
+
+      setVideoState((prev) => {
+        const state = [...prev];
+
+        state.forEach((rv) => {
+          // A video is uniquely identified by its name and storage type.
+          const match = toProtect.find(
+            (v) => v.videoName === rv.videoName && v.cloud === rv.cloud,
+          );
+
+          if (match) {
+            rv.isProtected = lock;
+          }
+        });
+
+        return state;
+      });
     };
 
     return (
@@ -146,7 +167,54 @@ export const populateDetailsCell = (
     );
   };
 
-  return <Box className="inline-flex">{renderProtectedIcon()}</Box>;
+  const renderTagIcon = () => {
+    const toTag = [video, ...video.multiPov];
+    const noPermission = !write && toTag.some((v) => v.cloud);
+
+    let tag = '';
+    let icon = <MessageSquare size={18} />;
+
+    let tooltip = noPermission
+      ? getLocalePhrase(language, Phrase.GuildNoPermission)
+      : getLocalePhrase(language, Phrase.TagButtonTooltip);
+
+    const foundTag = toTag.map((v) => v.tag).find((t) => t);
+
+    if (foundTag) {
+      tag = foundTag;
+      icon = <MessageSquareMore size={18} />;
+
+      if (tag.length > 50) {
+        tooltip = `${tag.slice(0, 50)}...`;
+      } else {
+        tooltip = tag;
+      }
+    }
+
+    return (
+      <Tooltip content={tooltip}>
+        <div onClick={(e) => e.stopPropagation()}>
+          <TagDialog
+            initialTag={tag}
+            videos={toTag}
+            setVideoState={setVideoState}
+            appState={appState}
+          >
+            <Button variant="ghost" size="xs" disabled={noPermission}>
+              {icon}
+            </Button>
+          </TagDialog>
+        </div>
+      </Tooltip>
+    );
+  };
+
+  return (
+    <Box className="inline-flex">
+      {renderProtectedIcon()}
+      {renderTagIcon()}
+    </Box>
+  );
 };
 
 export const populateLevelCell = (
@@ -250,7 +318,7 @@ export const populateViewpointCell = (
   };
 
   return (
-    <div className="flex ">
+    <div className="flex truncate">
       {renderSpecAndName()}
       {renderRemainingCount()}
     </div>